added password confirmation and file delete upon update

+++ resources/js/container/RegisterPage.js
@@ -8,12 +8,15 @@ import {
     InputLabel,
     Button,
     FormHelperText,
+    Alert,
+    Snackbar,
+    Slide,
 } from "@mui/material";
 import { makeStyles } from "@mui/styles";
 import axios from "axios";
 import React from "react";
 import swal from "sweetalert";
-import { useNavigate } from 'react-router-dom'
+import { useNavigate } from "react-router-dom";
 
 const useStyles = makeStyles((theme) => ({
     root: {
@@ -27,50 +30,65 @@ const useStyles = makeStyles((theme) => ({
 }));
 
 function RegisterPage() {
-    const history = useNavigate()
+    const history = useNavigate();
     const [register, setRegister] = React.useState({
         first_name: "",
         last_name: "",
         address: "",
         email: "",
         password: "",
-        number_phone: '',
+        number_phone: "",
         error_list: [],
     });
+    const [confirm, setConfirm] = React.useState("");
+    const [snackbarOpen, setSnackbarOpen] = React.useState(false)
 
     const handleInput = (e) => {
         e.persist();
         setRegister({ ...register, [e.target.name]: e.target.value });
     };
 
+    const handleConfirm = (e) => {
+        e.persist();
+        setConfirm(e.target.value);
+    };
+
+
+
     const registerSubmit = (e) => {
         e.preventDefault();
-        const data = {
-            first_name: register.first_name,
-            last_name: register.last_name,
-            address: register.address,
-            email: register.email,
-            password: register.password,
-            number_phone: register.number_phone,
-        };
+        if (register.password === confirm) {
+            const data = {
+                first_name: register.first_name,
+                last_name: register.last_name,
+                address: register.address,
+                email: register.email,
+                password: register.password,
+                number_phone: register.number_phone,
+            };
 
-        axios.get("/sanctum/csrf-cookie").then((response) => {
-            axios
-                .post('/api/register', data)
-                .then((res) => {
-                    if(res.data.status === 200) {
-                        localStorage.setItem('auth_token', res.data.token)
-                        localStorage.setItem('auth_firstName', res.data.first_name)
-                        localStorage.setItem('auth_lastName', res.data.last_name)
-                        localStorage.setItem('auth_email', res.data.email)
-                        swal('Success', res.data.message, "success")
-                        history('/')
+            axios.get("/sanctum/csrf-cookie").then((response) => {
+                axios.post("/api/register", data).then((res) => {
+                    if (res.data.status === 200) {
+                        localStorage.setItem("auth_token", res.data.token);
+                        swal("Success", res.data.message, "success");
+                        history("/");
                     } else {
-                        setRegister({...register, error_list: res.data.validation_errors})
+                        setRegister({
+                            ...register,
+                            error_list: res.data.validation_errors,
+                        });
                     }
                 });
-        });
+            });
+        } else {
+            setSnackbarOpen(true)
+        }
     };
+
+    const snackbarClose = () => {
+        setSnackbarOpen(false)
+    }
 
     const classes = useStyles();
     return (
@@ -100,13 +118,28 @@ function RegisterPage() {
                         component={"span"}
                         sx={{ color: "#FF674D", fontSize: "12px" }}
                     >
-                        Masuk Sekarang
+                        <Button
+                            onClick={() => history("/login")}
+                            variant="text"
+                        >
+                            <Typography fontSize="14px">
+                                Masuk Sekarang
+                            </Typography>
+                        </Button>
                     </Box>
                 </Typography>
                 <form onSubmit={registerSubmit}>
                     <Grid container sx={{ maxWidth: 450, mt: 2 }} spacing={2}>
                         <Grid item mobile={6}>
-                            <FormControl error={register.error_list.first_name?true:false} fullWidth variant="filled">
+                            <FormControl
+                                error={
+                                    register.error_list.first_name
+                                        ? true
+                                        : false
+                                }
+                                fullWidth
+                                variant="filled"
+                            >
                                 <InputLabel htmlFor="component-filled">
                                     First Name
                                 </InputLabel>
@@ -121,11 +154,19 @@ function RegisterPage() {
                                         input: classes.input,
                                     }}
                                 />
-                                <FormHelperText sx={{fontSize: 10}}>{register.error_list.first_name}</FormHelperText>
+                                <FormHelperText sx={{ fontSize: 10 }}>
+                                    {register.error_list.first_name}
+                                </FormHelperText>
                             </FormControl>
                         </Grid>
                         <Grid item mobile={6}>
-                            <FormControl error={register.error_list.last_name?true:false} fullWidth variant="filled">
+                            <FormControl
+                                error={
+                                    register.error_list.last_name ? true : false
+                                }
+                                fullWidth
+                                variant="filled"
+                            >
                                 <InputLabel htmlFor="component-filled">
                                     Last Name
                                 </InputLabel>
@@ -140,11 +181,19 @@ function RegisterPage() {
                                         input: classes.input,
                                     }}
                                 />
-                                <FormHelperText sx={{fontSize: 10}}>{register.error_list.last_name}</FormHelperText>
+                                <FormHelperText sx={{ fontSize: 10 }}>
+                                    {register.error_list.last_name}
+                                </FormHelperText>
                             </FormControl>
                         </Grid>
                         <Grid item mobile={12}>
-                            <FormControl error={register.error_list.address?true:false} fullWidth variant="filled">
+                            <FormControl
+                                error={
+                                    register.error_list.address ? true : false
+                                }
+                                fullWidth
+                                variant="filled"
+                            >
                                 <InputLabel htmlFor="component-filled">
                                     Address
                                 </InputLabel>
@@ -159,11 +208,21 @@ function RegisterPage() {
                                         input: classes.input,
                                     }}
                                 />
-                                <FormHelperText>{register.error_list.address}</FormHelperText>
+                                <FormHelperText>
+                                    {register.error_list.address}
+                                </FormHelperText>
                             </FormControl>
                         </Grid>
                         <Grid item mobile={12}>
-                            <FormControl error={register.error_list.number_phone?true:false} fullWidth variant="filled">
+                            <FormControl
+                                error={
+                                    register.error_list.number_phone
+                                        ? true
+                                        : false
+                                }
+                                fullWidth
+                                variant="filled"
+                            >
                                 <InputLabel htmlFor="component-filled">
                                     Phone Number
                                 </InputLabel>
@@ -178,11 +237,17 @@ function RegisterPage() {
                                         input: classes.input,
                                     }}
                                 />
-                                <FormHelperText>{register.error_list.number_phone}</FormHelperText>
+                                <FormHelperText>
+                                    {register.error_list.number_phone}
+                                </FormHelperText>
                             </FormControl>
                         </Grid>
                         <Grid item mobile={12}>
-                            <FormControl error={register.error_list.email?true:false} fullWidth variant="filled">
+                            <FormControl
+                                error={register.error_list.email ? true : false}
+                                fullWidth
+                                variant="filled"
+                            >
                                 <InputLabel htmlFor="component-filled">
                                     Email
                                 </InputLabel>
@@ -197,12 +262,20 @@ function RegisterPage() {
                                         input: classes.input,
                                     }}
                                 />
-                                <FormHelperText>{register.error_list.email}</FormHelperText>
+                                <FormHelperText>
+                                    {register.error_list.email}
+                                </FormHelperText>
                             </FormControl>
                         </Grid>
 
                         <Grid item mobile={12}>
-                            <FormControl error={register.error_list.password?true:false} fullWidth variant="filled">
+                            <FormControl
+                                error={
+                                    register.error_list.password ? true : false
+                                }
+                                fullWidth
+                                variant="filled"
+                            >
                                 <InputLabel htmlFor="component-filled">
                                     Password
                                 </InputLabel>
@@ -218,19 +291,28 @@ function RegisterPage() {
                                         input: classes.input,
                                     }}
                                 />
-                                <FormHelperText>{register.error_list.password}</FormHelperText>
+                                <FormHelperText>
+                                    {register.error_list.password}
+                                </FormHelperText>
                             </FormControl>
                         </Grid>
 
                         <Grid item mobile={12}>
-                            <FormControl error={register.error_list.password?true:false} fullWidth variant="filled">
+                            <FormControl
+                                error={
+                                    register.error_list.password ? true : false
+                                }
+                                fullWidth
+                                variant="filled"
+                            >
                                 <InputLabel htmlFor="component-filled">
                                     Confirm Password
                                 </InputLabel>
                                 <FilledInput
                                     id="component-filled"
                                     disableUnderline={true}
-                                    onChange={handleInput}
+                                    onChange={handleConfirm}
+                                    value={confirm}
                                     type="password"
                                     name="confirmPassword"
                                     classes={{
@@ -260,6 +342,16 @@ function RegisterPage() {
                     </Box>
                 </form>
             </Box>
+            <Snackbar anchorOrigin={{vertical: 'bottom', horizontal: 'center'} } open={snackbarOpen} autoHideDuration={6000} onClose={snackbarClose}>
+                <Alert
+                    onClose={snackbarClose}
+                    severity="error"
+                    variant="filled"
+                    sx={{ width: "100%" }}
+                >
+                    Password tidak sesuai!
+                </Alert>
+            </Snackbar>
         </Box>
     );
 }
